:art: split each document type into its own spec file
