+++ spec/data.rb
@@ -1,15 +1,12 @@
 # frozen_string_literal: true
 
 DATA_DIR = File.join(__dir__, 'data').freeze
-DIR_CUSTOM_V1 = File.join(DATA_DIR, 'custom', 'response_v1').freeze
-DIR_INVOICE_V4 = File.join(DATA_DIR, 'invoice', 'response_v4').freeze
-DIR_RECEIPT_V4 = File.join(DATA_DIR, 'receipt', 'response_v4').freeze
-DIR_FINANCIAL_DOCUMENT_V1 = File.join(DATA_DIR, 'financial_document', 'response_v1').freeze
-DIR_PROOF_OF_ADDRESS_V1 = File.join(DATA_DIR, 'proof_of_address', 'response_v1').freeze
-DIR_PASSPORT_V1 = File.join(DATA_DIR, 'passport', 'response_v1').freeze
-DIR_EU_LICENSE_PLATE_V1 = File.join(DATA_DIR, 'eu', 'license_plate', 'response_v1').freeze
-DIR_SHIPPING_CONTAINER_V1 = File.join(DATA_DIR, 'shipping_container', 'response_v1').freeze
-DIR_US_BANK_CHECK_V1 = File.join(DATA_DIR, 'us', 'bank_check', 'response_v1').freeze
-DIR_FR_BANK_ACCOUNT_DETAILS_V1 = File.join(DATA_DIR, 'fr', 'bank_account_details', 'response_v1').freeze
-DIR_FR_CARTE_VITALE_V1 = File.join(DATA_DIR, 'fr', 'carte_vitale', 'response_v1').freeze
-DIR_FR_ID_CARD_V1 = File.join(DATA_DIR, 'fr', 'id_card', 'response_v1').freeze
+
+def load_json(dir_path, file_name)
+  file_data = File.read(File.join(dir_path, file_name))
+  JSON.parse(file_data)
+end
+
+def read_file(dir_path, file_name)
+  File.read(File.join(dir_path, file_name)).strip
+end
